refactor: extract isoDate
